Use the FindQjson.cmake file made by pinotree for chokoq because it works. The old one didn't (for me).
It did set QJSON_LIBRARIES=qjson:QtCore which meant the qjson lib was not
found on my system because noone added the -L<path> thing.

Pinos version sets QJSON_LIBRARY=/full/path/to/libjson but does not add
anything about QtCore. Still it works.

This is the second time we fix problems with finding qjson. There is still
a problem with FindQjson.cmake against FindQJSON.cmake on windows if one is
not very careful.

CCMAIL: kde-buildsystem@kde.org